+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.0.0 [February 23, 2025]
 
 - **BREAKING CHANGE**: Updated default host to our new synchronous API subdomain: sync.api.docspring.com. (EU customers should use sync.api-eu.docspring.com). Removed all custom polling code from library since this logic is now handled by the API service running on our sync subdomain
+- Fixed User-Agent header: docspring-elixir-x.x.x
 
 ### 1.0.0 [Jan 30, 2025]